cleanup(core): cleanup lock file functions
